API Pagamentos - Exceptions

+++ pagamentosAPI/src/main/java/com/totalshakes/pagamentos/service/PagamentoServiceImplementation.java
@@ -1,14 +1,14 @@
 package com.totalshakes.pagamentos.service;
 
 import com.totalshakes.pagamentos.common.CommonExtensions;
-import com.totalshakes.pagamentos.model.Pagamento;
 import com.totalshakes.pagamentos.dto.PagamentoDTO;
+import com.totalshakes.pagamentos.exceptions.NaoHaPagamentosException;
+import com.totalshakes.pagamentos.exceptions.PagamentoNaoEncontradoException;
+import com.totalshakes.pagamentos.model.Pagamento;
 import com.totalshakes.pagamentos.repository.PagamentoRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-
-//TODO - FIX: updatePagamento
 
 @Service
 public class PagamentoServiceImplementation extends CommonExtensions implements PagamentoService {
@@ -26,21 +26,25 @@ public class PagamentoServiceImplementation extends CommonExtensions implements 
     }
 
     @Override
-    public PagamentoDTO findPagamentoById(long id) {
+    public PagamentoDTO findPagamentoById(long id) throws PagamentoNaoEncontradoException {
+        if (!pagamentoRepository.findById(id).isPresent())
+            throw new PagamentoNaoEncontradoException();
 
-        PagamentoDTO pagamentoEncontrado = super.convertToDTO(pagamentoRepository.findById(id).get(),
+        return super.convertToDTO(pagamentoRepository.findById(id).get(),
                 PagamentoDTO.class);
-        return pagamentoEncontrado;
     }
 
     @Override
-    public List<Pagamento> findAllPagamentos() {
+    public List<Pagamento> findAllPagamentos() throws NaoHaPagamentosException {
+        if (pagamentoRepository.findAll().isEmpty())
+            throw new NaoHaPagamentosException();
 
         List<Pagamento> listaPagamentos = pagamentoRepository.findAll();
         listaPagamentos.stream().map(pagamento -> (super.convertToDTO(pagamento, PagamentoDTO.class)));
 
         return listaPagamentos;
     }
+
     @Override
     public void updatePagamento(long id, PagamentoDTO pagamentoDTO) {
 
@@ -49,8 +53,9 @@ public class PagamentoServiceImplementation extends CommonExtensions implements 
     }
 
     @Override
-    public void deletePagamentoById(long id) {
-
+    public void deletePagamentoById(long id) throws PagamentoNaoEncontradoException {
+        if (!pagamentoRepository.findById(id).isPresent())
+            throw new PagamentoNaoEncontradoException();
         pagamentoRepository.deleteById(id);
     }
 }
